+++ apps/docs/app/(home)/blog/[slug]/page.tsx
@@ -3,8 +3,8 @@ import Link from "next/link";
 import { blog, BlogPage } from "@/app/source";
 import { buttonVariants } from "@/components/ui/button";
 import Image from "next/image";
-
 import profilePic from "../../../../components/testimonials/profiles/Mc0m3zkD_400x400.jpg";
+import { useMDXComponents } from "@/mdx-components";
 
 interface Param {
   slug: string;
@@ -16,10 +16,9 @@ export default function Page({
   params: Param;
 }): React.ReactElement {
   const page = blog.getPage([params.slug]) as BlogPage;
+  const mdxComponents = useMDXComponents({});
 
   if (!page) notFound();
-
-  const MDX = (page.data as any).exports.default;
 
   return (
     <main className="px-4">
@@ -30,13 +29,15 @@ export default function Page({
         >
           Back
         </Link>
-        <p className="text-xs text-gray-500">
-          {(page.data.date as Date).toLocaleString("en-US", {
-            year: "numeric",
-            month: "2-digit",
-            day: "2-digit",
-          })}
-        </p>
+        {!!page.data.date && (
+          <p className="text-xs text-gray-500">
+            {(page.data.date as Date).toLocaleString("en-US", {
+              year: "numeric",
+              month: "2-digit",
+              day: "2-digit",
+            })}
+          </p>
+        )}
       </div>
       <div
         className="mx-auto w-full max-w-screen-xl rounded-xl border py-12 md:px-8"
@@ -60,7 +61,7 @@ export default function Page({
         </div>
       </div>
       <article className="prose lg:prose-lg mx-auto w-full max-w-screen-sm py-8">
-        <MDX />
+        <page.data.body components={mdxComponents} />
       </article>
       <div className="mx-auto mb-20 flex w-full max-w-screen-sm items-start gap-3">
         <Image