+++ packages/react-langgraph/src/index.ts
@@ -1,4 +1,5 @@
-export { useLangChainLangGraphRuntime } from "./useLangGraphRuntime";
+export { useLangGraphRuntime } from "./useLangGraphRuntime";
+
 export { useLangGraphMessages } from "./useLangGraphMessages";
 export { convertLangchainMessages } from "./convertLangchainMessages";
 export type {