+++ packages/react/src/primitive-hooks/composer/useComposerAddAttachment.tsx
@@ -1,16 +1,21 @@
 import { useCallback } from "react";
-import { useThreadContext } from "../../context";
+import { useComposer } from "../../context";
+import { useThreadComposerStore } from "../../context/react/ThreadContext";
 
 export const useComposerAddAttachment = () => {
-  const { useComposer } = useThreadContext();
-
   const disabled = useComposer((c) => !c.isEditing);
 
+  const threadComposerStore = useThreadComposerStore();
+  const threadRuntimeStore = useThreadComposerStore();
   const callback = useCallback(() => {
-    const { addAttachment } = useComposer.getState();
+    const { addAttachment } = threadComposerStore.getState();
+    const { attachmentAccept } = threadRuntimeStore.getState();
 
     const input = document.createElement("input");
     input.type = "file";
+    if (attachmentAccept !== "*") {
+      input.accept = attachmentAccept;
+    }
 
     input.onchange = (e) => {
       const file = (e.target as HTMLInputElement).files?.[0];
@@ -19,7 +24,7 @@ export const useComposerAddAttachment = () => {
     };
 
     input.click();
-  }, [useComposer]);
+  }, [threadComposerStore, threadRuntimeStore]);
 
   if (disabled) return null;
   return callback;