+++ packages/react-langgraph/CHANGELOG.md
@@ -1,5 +1,62 @@
 # @assistant-ui/react-langgraph
 
+## 0.0.8
+
+### Patch Changes
+
+- 155d6e7: chore: update dependencies
+- Updated dependencies [926dce5]
+- Updated dependencies [155d6e7]
+- Updated dependencies [f80226f]
+  - @assistant-ui/react@0.5.60
+
+## 0.0.7
+
+### Patch Changes
+
+- e4863bb: feat(langgraph): add support for switching threads
+- Updated dependencies [e4863bb]
+- Updated dependencies [e4863bb]
+  - @assistant-ui/react@0.5.56
+
+## 0.0.6
+
+### Patch Changes
+
+- c348553: chore: update dependencies
+- Updated dependencies [0f99aa6]
+- Updated dependencies [c348553]
+  - @assistant-ui/react@0.5.54
+
+## 0.0.5
+
+### Patch Changes
+
+- 934758b: feat: automatically cancel tool calls if user sends a new message
+
+## 0.0.4
+
+### Patch Changes
+
+- 184d836: feat: allow multiple message sends to support pending tool call cancellations
+
+## 0.0.3
+
+### Patch Changes
+
+- c1c0440: refactor: rename to useLangGraphRuntime
+- Updated dependencies [164e46c]
+- Updated dependencies [5eccae7]
+  - @assistant-ui/react@0.5.51
+
+## 0.0.2
+
+### Patch Changes
+
+- 04f6fc8: chore: update deps
+- Updated dependencies [04f6fc8]
+  - @assistant-ui/react@0.5.50
+
 ## 0.0.1
 
 ### Patch Changes